fix: delete legacy script & reorgnize old script
