--- qlib_data/features/data.csv
@@ -1,6 +1,0 @@
-instrument,datetime,$open,$high,$low,$close,$volume
-test,2023-01-01,150.0,155.0,149.0,154.0,1000000
-test,2023-01-02,154.0,158.0,153.0,157.0,1200000
-test,2023-01-03,157.0,160.0,156.0,159.0,900000
-test,2023-01-04,159.0,162.0,158.0,161.0,1100000
-test,2023-01-05,161.0,163.0,160.0,162.0,800000
