--- qlib_data/instruments/all.txt
@@ -1,1 +1,0 @@
-test
