--- qlib_data/validation_report.txt
@@ -1,15 +1,0 @@
-=== QLIB 数据验证报告 ===
-
-验证时间: 2025-07-02 14:34:35.293384
-数据目录: qlib_data
-
-总体状态: ✓ 通过
-错误数量: 0
-警告数量: 0
-
-=== 数据统计 ===
-总记录数: 5
-股票数量: 1
-日期范围: 2023-01-01T00:00:00 至 2023-01-05T00:00:00
-交易日数: 5
-
