--- raw_data/test_data_from_csv.csv
@@ -1,6 +1,0 @@
-date,open,high,low,close,volume
-2023-01-01,150.0,155.0,149.0,154.0,1000000
-2023-01-02,154.0,158.0,153.0,157.0,1200000
-2023-01-03,157.0,160.0,156.0,159.0,900000
-2023-01-04,159.0,162.0,158.0,161.0,1100000
-2023-01-05,161.0,163.0,160.0,162.0,800000
